add Registers::pc method
If #29 is ever resolved, a similar method could be added to `RegId`,
enabling more efficient `pc` queries. Note that the less efficient
approach would still need to be supported, as the ability to query
individual registers is not a required feature implemented by all
targets.

+++ src/arch/msp430/mod.rs
@@ -15,7 +15,7 @@ pub enum Msp430<RegIdImpl: RegId = reg::id::Msp430RegId> {
 }
 
 impl<RegIdImpl: RegId> Arch for Msp430<RegIdImpl> {
-    type Usize = u32;
+    type Usize = u16;
     type Registers = reg::Msp430Regs;
     type RegId = RegIdImpl;
     type BreakpointKind = usize;
